+++ tests/functor_application.ml
@@ -1,3 +1,4 @@
+(** Some documentation *)
 module type Source = sig
   type t
   val x : t